python/agents/auto-insurance-agent/auto_insurance_agent/requirements.txt
google-adk
google-cloud-secret-manager